Added getRunningActivity; Precautions in InstrumentationDriver.getRootElement
There is now a ActivityUtils class with a getRunningActivity() method, which
depends on the overridden callActivityOnResume() and callActivityOnPause()
methods in TestRunner to maintain a currentActivity.

InstrumentationDriver.getRootElement now throws an exception if the test
activity isn't started and uses runningActivity.getWindow().getDecorView()
as the root element (used for finding elements).

Change-Id: Ib43ed5b2b8f17bc2dc20f21c24d76afe54fa34cd