Fixed issue with clicking on partially visible elements.
Because UiElement.click() was performed on the center of elements, it didn't
work when the element's center was off screen. To fix this, I added a
getVisibleBounds() method to UiElement, which is implemented by ViewElement
and UiAutomationElement. This method is used in ClickAction to make sure that
the click is performed on the visible portion of the element.

DeviceUtils was added with getDisplayBounds().

Lastly, UiAutomationContext now holds Instrumentation instead of UiAutomation,
which also forced other files to change accordingly.
Change-Id: Ic2fd163a9759a0d95bd413b557976ee8a3d044f9