Fixed ViewElement.getVisibleBounds()
Previously, getVisibleBounds was only getting the coordinates of an
element relative to the topmost view, which is not the screen in the
case that there is a pop-up dialog. By using view.getLocationOnScreen,
we can adjust the relative coordinates to apply to the entire screen and
have the driver click on the correct place.

Bug: b/9743953
Change-Id: I51a2f7fc8974fba9d08c660b6e7926eb0e305abb