fix setText and remove replaceText
setText was a misleading method because it did not clear existing text,
which was implied by the name.
This change first puts the UiElement in focus, then clears existing text,
finally sets the new text.
This change is not backward compatible in many cases, so please modify your
client code then we can coordinate to submit at the same time.

Change-Id: I3a959301734db9075a768b1d63019af8320dc89b