Fix the remaining JavaDoc warnings
We should have a consistent style for JavaDoc, i.e. @param for all
parameters of a method, or none. And a consistent format. I used
"Reformat code" from Android Studio on some files I touched in this cl.
Now the consistency is only on a per-file basis. We may apply it
incrementally.

Change-Id: Icfe2ef223d93e78e42b400a1dd2b50a0ef19a13a