Change minSdkVersion to 8
Guard access to newer API.
Ideally we should have lint check the usage of our API in user code.
E.g. the users specify minSdkVersion of their app to a value lower
than 11, but use SingleKeyAction.CTRL_MOVE_HOME.
It should be a feature of the Android lint, as described in
http://stackoverflow.com/questions/14066857/sdk-version-required-annotation
We may write our own lint extension, but it's hard to get it into
users' workflow.

Change-Id: I8efe88475646530dd8c90ccf443fb3ee9b62c900